+++ tests/conftest.py
@@ -1,0 +1,3 @@
+import os
+
+os.environ.setdefault("VULNORAIQ_ALLOW_TEST_FIXTURE_TARGETS", "true")